Replace deprecated template_file with filetemplate
The old way did not work when terraform was run from a windows machine, the yml files were not created. Plus, CRLF were printed as-is in the yml files when run the commands manually. Finally, the template_file is deprecated. Therefore, I am implementing `local_file` with `templatefile` to solve those issues in one go.